+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
   * Hide sesitive information when printing data sent to OSM in debug mode
   * Add archived attribute to Event
+  * Add :include\_archived option to Api.get\_events method
 
 ## Version 0.0.19